+++ Pic32/globals.h
@@ -34,7 +34,7 @@ extern volatile FAULT_TYPE fault;
 #ifndef WCM_PRESENT
     extern volatile bool sendHeartbeat;
 #endif
-extern unsigned int loopIteration;
+extern unsigned long long loopIteration;
 /******************************************************************************/
